test: split simulator_test.go into focused per-endpoint test files
